REF: Refactor code
