+++ ms-bank-svc/src/main/java/com/anasdidi/msbanksvc/common/BaseRoute.java
@@ -4,9 +4,9 @@ import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.RoutingContext;
 
-public abstract class BaseHandler {
+public abstract class BaseRoute {
 
-  protected abstract void handle(RoutingContext ctx);
+  protected abstract JsonObject process(RoutingContext ctx);
 
   protected abstract HttpMethod getHttpMethod();
 
@@ -18,7 +18,12 @@ public abstract class BaseHandler {
     ctx.response().end(resBody.encode());
   }
 
-  protected BaseDTO getDTO(RoutingContext ctx) {
+  protected final BaseDTO getDTO(RoutingContext ctx) {
     return ctx.get(Constants.Context.DTO);
+  }
+
+  protected final void handle(RoutingContext ctx) {
+    JsonObject responseBody = process(ctx);
+    ctx.response().end(responseBody.encode());
   }
 }
